requirements.txt
flask>=3.0.0
opencv-python-headless>=4.5.0
pytesseract>=0.3.10
pillow>=9.0.0
numpy>=1.20.0
